# api/swagger.yml
swagger: "2.0"
info:
  description: Provide product recommendation and product pricing service
  version: "1.0"
  title: DataTeam - DEMO API
consumes:
  - application/json
produces:
  - application/json

#basePath: /

# Paths supported by the server application
paths:
...
  /getTime:
    get:
      operationId: main.get_timestamp
      tags:
        - "[Demo]"
      summary: getServerTime
      responses:
        200:
          description: Successfully query customer product price
        400:
          description: Not Found
...